Move ContainerEndpoint from configserver => config-model-api
